bingo - for posterity's sake
There was a party once, where Dan said to me "We should make a bingo card generator."  So, we did.  This is the random-number stuffing part, but unfortunately the other half is missing.  There was a postscript file on dan's ipod called "bingo.ps".  If anyone happens to have a copy, please send it to me, and/or import it into this repository.